fix: para kolonlarini BigInteger yap (int4 tasmasi)
Istanbul'da ilan fiyati int4 sinirini (2,1 milyar TL) asabiliyor; yeni veride
8,75 milyar TL'lik bir ilan vardi. SQLAlchemy toplu insert yaptigindan bu tek
satir 1000'lik partinin tamamini dusuruyor ve ingest komple geri aliniyordu
(CSV doluydu ama DB'ye hicbir sey girmiyordu).

listings.fiyat ve valuations tutar kolonlari (tahmini_fiyat, birim_m2_fiyat,
fiyat_alt, fiyat_ust, beklenen_fiyat) BigInteger'a cevrildi + Alembic migration
(a1f2c3d4e5f6). Migration uygulandi, ingest yeniden calisti: 1841 yeni ilan
eklendi, atlanan 0.

+++ app/models/listing.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import String, Integer, Float, Boolean, DateTime, func
+from sqlalchemy import String, Integer, BigInteger, Float, Boolean, DateTime, func
 from sqlalchemy import Enum as SqlEnum
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -19,7 +19,8 @@ class Listing(Base):
     ilce: Mapped[str] = mapped_column(String(50), index=True)
     mahalle: Mapped[str | None] = mapped_column(String(100), nullable=True)
 
-    fiyat: Mapped[int] = mapped_column(Integer, index=True)
+    # BigInteger: Istanbul'da fiyat 2,1 milyar TL'yi (int4 siniri) asabiliyor.
+    fiyat: Mapped[int] = mapped_column(BigInteger, index=True)
     brut_metrekare: Mapped[int] = mapped_column(Integer)
     net_metrekare: Mapped[int] = mapped_column(Integer)
     oda_sayisi: Mapped[str] = mapped_column(String(10))
