+++ app/models/valuation.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import String, Integer, DateTime, func
+from sqlalchemy import String, Integer, BigInteger, DateTime, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.database import Base
@@ -30,15 +30,16 @@ class Valuation(Base):
     esya_durumu: Mapped[str | None] = mapped_column(String(50), nullable=True)
     aidat: Mapped[int] = mapped_column(Integer, default=0)
 
-    # Tahmin sonucu.
-    tahmini_fiyat: Mapped[int] = mapped_column(Integer)
-    birim_m2_fiyat: Mapped[int] = mapped_column(Integer)
-    fiyat_alt: Mapped[int] = mapped_column(Integer)
-    fiyat_ust: Mapped[int] = mapped_column(Integer)
+    # Tahmin sonucu. BigInteger: pahali konutlarda tutarlar 2,1 milyar TL'yi
+    # (int4 siniri) asabilir.
+    tahmini_fiyat: Mapped[int] = mapped_column(BigInteger)
+    birim_m2_fiyat: Mapped[int] = mapped_column(BigInteger)
+    fiyat_alt: Mapped[int] = mapped_column(BigInteger)
+    fiyat_ust: Mapped[int] = mapped_column(BigInteger)
 
     # Kullanici bir beklenen fiyat girdiyse: fiyat + yorum ("uygun"/"normal"/
     # "pahali"). Girmediyse null kalir.
-    beklenen_fiyat: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    beklenen_fiyat: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     yorum_durum: Mapped[str | None] = mapped_column(String(20), nullable=True)
 
     olusturulma_tarihi: Mapped[datetime] = mapped_column(
